+++ src/main/scala/com/atomist/tree/utils/NodeUtils.scala
@@ -1,6 +1,7 @@
 package com.atomist.tree.utils
 
 import com.atomist.graph.GraphNode
+import com.atomist.rug.kind.grammar.ParsedNode
 import com.atomist.tree.TreeNode
 import com.atomist.tree.content.text.PositionedTreeNode
 import org.springframework.util.ReflectionUtils
@@ -62,8 +63,8 @@ object NodeUtils {
     case _ => ""
   }
 
-  def positionedValue(gn: GraphNode, in: String): String = gn match {
-    case pn: PositionedTreeNode => in.substring(pn.startPosition.offset, pn.endPosition.offset)
+  def positionedValue(gn: Object, in: String): String = gn match {
+    case pn: ParsedNode => in.substring(pn.startOffset, pn.endOffset)
     case _ => ???
   }